Support implementations of JAX functions within a context.
The idea here is to support overriding the implementation of higher order JAX
functions within a limited scope, for libraries such as Haiku and Flax that
implement their own versions of this functions that support mutation.

Usage example:

    with jax.override_context({'lax.scan': my_scan}):
      # all calls to lax.scan() are replaced by my_scan()
      ...

Ultimately, it would be great to replace this with a unified interface for
mutable state in JAX, but this could be a convenient temporary measure.